Rename and move logprep.output.writing_output to logprep.connector.jsonl.output
